Create shared OOXML writer file.
This is for functions used by both Powerpoint and Docx writers.
